minor refactor – ecs package organization
