+++ Showcases/geo_voyage/app/src/main/java/com/meta/pixelandtexel/geovoyage/activities/MainActivity.kt
@@ -14,12 +14,12 @@ import com.meta.pixelandtexel.geovoyage.R
 import com.meta.pixelandtexel.geovoyage.Spin
 import com.meta.pixelandtexel.geovoyage.Spinnable
 import com.meta.pixelandtexel.geovoyage.Tether
-import com.meta.pixelandtexel.geovoyage.ecs.grabbablenorotation.GrabbableNoRotationSystem
-import com.meta.pixelandtexel.geovoyage.ecs.landmarkspawn.LandmarkSpawnSystem
-import com.meta.pixelandtexel.geovoyage.ecs.pinnable.PinnableSystem
-import com.meta.pixelandtexel.geovoyage.ecs.spin.SpinSystem
-import com.meta.pixelandtexel.geovoyage.ecs.spinnable.SpinnableSystem
-import com.meta.pixelandtexel.geovoyage.ecs.tether.TetherSystem
+import com.meta.pixelandtexel.geovoyage.ecs.GrabbableNoRotationSystem
+import com.meta.pixelandtexel.geovoyage.ecs.LandmarkSpawnSystem
+import com.meta.pixelandtexel.geovoyage.ecs.PinnableSystem
+import com.meta.pixelandtexel.geovoyage.ecs.SpinSystem
+import com.meta.pixelandtexel.geovoyage.ecs.SpinnableSystem
+import com.meta.pixelandtexel.geovoyage.ecs.TetherSystem
 import com.meta.pixelandtexel.geovoyage.enums.PlayMode
 import com.meta.pixelandtexel.geovoyage.enums.SettingsKey
 import com.meta.pixelandtexel.geovoyage.models.GeoCoordinates
