+++ Showcases/geo_voyage/app/src/main/java/com/meta/pixelandtexel/geovoyage/ecs/LandmarkSpawnSystem.kt
@@ -1,6 +1,6 @@
 // (c) Meta Platforms, Inc. and affiliates. Confidential and proprietary.
 
-package com.meta.pixelandtexel.geovoyage.ecs.landmarkspawn
+package com.meta.pixelandtexel.geovoyage.ecs
 
 import android.content.Context
 import android.net.Uri
