+++ Showcases/geo_voyage/app/src/main/java/com/meta/pixelandtexel/geovoyage/ecs/SpinSystem.kt
@@ -1,6 +1,6 @@
 // (c) Meta Platforms, Inc. and affiliates. Confidential and proprietary.
 
-package com.meta.pixelandtexel.geovoyage.ecs.spin
+package com.meta.pixelandtexel.geovoyage.ecs
 
 import com.meta.pixelandtexel.geovoyage.Spin
 import com.meta.pixelandtexel.geovoyage.utils.MathUtils.fromAxisAngle
